Added HMAC validator Reverted streams to support java 7
