+++ src/main/java/com/adyen/Util/Util.java
@@ -1,22 +1,15 @@
 package com.adyen.Util;
 
 import com.adyen.model.Amount;
-import com.google.common.io.BaseEncoding;
 
-import javax.crypto.Mac;
-import javax.crypto.spec.SecretKeySpec;
 import java.math.BigDecimal;
-import java.nio.charset.Charset;
-import java.security.SignatureException;
 import java.text.SimpleDateFormat;
-import java.util.*;
-import java.util.stream.Collectors;
-import java.util.stream.Stream;
+import java.util.Calendar;
+import java.util.Date;
+import java.util.Iterator;
+import java.util.List;
 
 public class Util {
-    private final static String HMAC_SHA256_ALGORITHM = "HmacSHA256";
-    private final static Charset C_UTF8 = Charset.forName("UTF8");
-
     public static <T> String implode(String glue, List<T> list) {
         // list is empty, return empty string
         if (list == null || list.isEmpty()) {
@@ -67,45 +60,6 @@ public class Util {
             default:
                 return 2;
         }
-    }
-
-    // To calculate the HMAC SHA-256
-    public static String calculateHMAC(String data, String key) throws java.security.SignatureException {
-        try {
-            byte[] rawKey = BaseEncoding.base16().decode(key);
-            // Create an hmac_sha256 key from the raw key bytes
-            SecretKeySpec signingKey = new SecretKeySpec(rawKey, HMAC_SHA256_ALGORITHM);
-
-            // Get an hmac_sha256 Mac instance and initialize with the signing
-            // key
-            Mac mac = Mac.getInstance(HMAC_SHA256_ALGORITHM);
-
-            mac.init(signingKey);
-
-            // Compute the hmac on input data bytes
-            byte[] rawHmac = mac.doFinal(data.getBytes(C_UTF8));
-
-            // Base64-encode the hmac
-            return BaseEncoding.base64().encode(rawHmac);
-
-        } catch (Exception e) {
-            throw new SignatureException("Failed to generate HMAC : " + e.getMessage());
-        }
-    }
-
-    private static String escapeVal(String val) {
-        if (val == null) {
-            return "";
-        }
-        return val.replace("\\", "\\\\").replace(":", "\\:");
-    }
-
-    public static String getDataToSign(SortedMap<String, String> postParameters) {
-        String signedData = Stream.concat(postParameters.keySet().stream(), postParameters.values().stream())
-                .map(Util::escapeVal)
-                .collect(Collectors.joining(":"));
-
-        return signedData;
     }
 
     /**
